--- common/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "common"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-mockall = "0.11.4"
-async-trait = "0.1.72"
-
-
-[features]
-initiator = []
-repeater = []
-responder = []
