--- common/README.md
@@ -1,6 +1,0 @@
-# Common
-
-## Introduction
-This implements cross component datastructures and messages
-
-##
