--- common/src/lib.rs
@@ -1,14 +1,0 @@
-pub mod models {
-    pub mod application_request_format;
-    pub mod link_allocation_policy;
-    pub mod performance_indicator;
-    pub mod resources;
-}
-
-pub mod interfaces {
-    pub mod connection_manager;
-    pub mod hardware_monitor;
-    pub mod realtime_controller;
-    pub mod routing_daemon;
-    pub mod rule_engine;
-}
