--- connection_manager/Settings.toml
@@ -1,3 +1,0 @@
-num_thread = 10
-port = 52244
-host = "0.0.0.0"
