--- hardware_monitor/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod hardware_monitor;
-mod link_resource_table;
