--- hardware_monitor/src/link_resource_table.rs
@@ -1,12 +1,0 @@
-/// Record generated link resource
-pub(crate) struct LinkResourceTable {}
-
-impl LinkResourceTable {
-    pub fn new() {}
-
-    pub fn push_new_resource() {}
-
-    pub fn pop_resource() {}
-
-    pub fn transfer_ownership() {}
-}
