+++ qrsa/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "connection_manager"
+name = "qrsa"
 version = "0.1.0"
 edition = "2021"
 
@@ -11,13 +11,4 @@ config = {version = "0.13.3", features = ["toml"]}
 tokio = {version = "1.29.1", features = ["full"]}
 uuid = {version = "1.4.1", features = ["v4"]}
 async-trait = "0.1.72"
-
-common = {path = "../common"}
-
-[dev-dependencies]
 mockall = "0.11.4"
-
-[features]
-initiator = []
-repeater = []
-responder = []
