+++ qrsa/src/common/link_allocation_policy.rs
