+++ qrsa/src/common/performance_indicator.rs
