+++ qrsa/src/common/resources.rs
