+++ qrsa/src/common/ruleset.rs
@@ -1,0 +1,9 @@
+/// Actual RuleSet contents that contains a set of instructions for repeaters
+///
+pub struct RuleSet {}
+
+impl RuleSet {
+    pub fn new() {}
+
+    pub fn add_instruction() {}
+}
