+++ qrsa/src/connection_manager/README.md
