+++ qrsa/src/connection_manager/error.rs
