+++ qrsa/src/hardware_monitor/README.md
