+++ qrsa/src/hardware_monitor/hardware_monitor.rs
