+++ qrsa/src/hardware_monitor/interface.rs
