+++ qrsa/src/realtime_controller/README.md
