+++ qrsa/src/realtime_controller/hal_wrapper.rs
