+++ qrsa/src/rule_engine/rule_engine.rs
@@ -1,6 +1,6 @@
-use crate::la_policy_manager::LinkAllocationPolicyManager;
-use crate::resource_allocator::ResourceAllocator;
-use crate::scheduler::RuleSetScheduler;
+use super::la_policy_manager::LinkAllocationPolicyManager;
+use super::resource_allocator::ResourceAllocator;
+use super::scheduler::RuleSetScheduler;
 
 /// A component that executes RuleSet
 pub struct RuleEngine {
