--- realtime_controller/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "realtime_controller"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-
-
-[features]
-# Simulation mode
-simulation = []
-# Deployment mode for real device
-deploy = []
