--- realtime_controller/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod hal_wrapper;
-pub mod instruction_converter;
-pub mod realtime_controller;
