--- routing_daemon/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "routing_daemon"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-
-[features]
-simulation = []
-deploy = []
