--- rule_engine/src/lib.rs
@@ -1,8 +1,0 @@
-pub mod rule_engine;
-
-mod la_policy_manager;
-mod resource_allocator;
-mod scheduler;
-
-// ruleet runtime modules
-pub(crate) mod runtime;
